+++ ToSimple/src/main/java/model/QuestionnaireSpider.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import com.google.gson.Gson;
 
+import net.sf.json.JSONObject;
+
 
 
 public class QuestionnaireSpider {
@@ -13,6 +15,7 @@ public class QuestionnaireSpider {
 	
 	public static class Question{
 		public String questionTitle;
+		public boolean forced;
 		public int type;
 		public boolean mix;
 		public List<Choice> choices;
@@ -20,13 +23,25 @@ public class QuestionnaireSpider {
 	
 	public static class Choice{
 		public String text;
-		public boolean withBlank;
+		//public boolean withBlank;
 	}
 	
 	public static String toJson(QuestionnaireSpider q){
 		Gson gson = new Gson();
         String json = gson.toJson(q);
         return json;
+	}
+	
+	public static Questionnaire toQuestionnaire(QuestionnaireSpider q){
+		Gson gson = new Gson();
+        String json = gson.toJson(q);
+        JSONObject jobj=JSONObject.fromObject(json);
+        jobj.put("status", 0);
+        jobj.put("answerNumber", 0);
+        jobj.put("authorId", 1);
+        jobj.put("createDate", "1900-01-01T00:00:00.000Z");
+        
+        return new Questionnaire(jobj.toString());
 	}
