+++ ToSimple/src/main/java/service/QuestionnaireService.java
@@ -69,5 +69,10 @@ public interface QuestionnaireService {
 	public List<Questionnaire> getReportedQuestionnaire();
 	
 	public List<Questionnaire> getReportedQuestionnaireByPage(int page , int pageSize,CountUtils countUtils);
+	
+	public Questionnaire getSojumpQuestionnaire(String sojumpId);
+	
+	public int saveSojumpQuestionnaire(String sojumpId);
+	
 
 }
